+++ src/api/notice/notice-kind.ts
@@ -1,0 +1,6 @@
+export enum NoticeKind {
+  RECRUIT = 'recruit',
+  EVENT = 'event',
+  NORMAL = 'general',
+  ACADEMIC = 'academic',
+}
