+++ src/app/components/molecules/searchTagSelect/SearchTagSelect.tsx
@@ -1,4 +1,4 @@
-import { NoticeKind } from '@/api/notice/notice';
+import { NoticeKind } from '@/api/notice/notice-kind';
 
 import Button from '../../atoms/Button';
 import Chip from '../Chip';
